move rpauthz to rm
